+++ Statistic.Domain/QuizStatistic.cs
@@ -9,7 +9,7 @@ namespace Statistic.Domain
         public int TotalAnswersCount { get; set; }
         public int CorrectAnswersCount { get; set; }
         public int FailedAnswersCount { get; set; }
-        public double CorrectPercent { get; set; }
+        public decimal CorrectPercent { get; set; }
         public Guid UserId { get; set; }
         public string UserName { get; set; }
         public DateTime TimeStamp { get; set; }
