+++ Statistic.Domain/UserStatistic.cs
@@ -9,7 +9,7 @@ namespace Statistic.Domain
         public int QuestionsCount { get; set; }
         public int CorrectAnswersCount { get; set; }
         public int FailedAnswersCount { get; set; }
-        public double CorrectPercent { get; set; }
+        public decimal CorrectPercent { get; set; }
         public Guid QuizId { get; set; }
         public string QuizTitle { get; set; }
         public DateTime TimeStamp { get; set; }
